Fix: 프론트엔드 - 다대구 연동 로그인 고객만 가능하도록 설정

+++ frontend/src/components/atoms/didLoginButton.tsx
@@ -38,7 +38,11 @@ const DidLoginButton = ({ storeNum, tableNumber }: DidLoginButtonProps) => {
                     }
                 );
 
-                console.log("response", response.data);
+                const userRole = response.data.auth.role;
+                if (userRole != "user") {
+                    alert("고객만 연동 로그인이 가능합니다.");
+                    return;
+                }
 
                 localStorage.clear();
                 localStorage.setItem("tokenType", response.data.tokenType);
